fix(frontend): Borrar mascota
Ya funciona pero falta que se actualize la info en pantalla

+++ frontend/src/pages/userProfile/UserProfile.js
@@ -72,38 +72,37 @@ const WrapperListadoCards = styled.div`
 `;
 
 const UserProfile = () => {
-  const { userName, token } = JSON.parse(localStorage.getItem('token'));
+  const { token } = JSON.parse(localStorage.getItem('token'));
 
   const userMe = useFetch('http://localhost:8000/api/users/me', token);
-  console.log(userMe);
 
   const MascotasRegistradas = useFetch('http://localhost:8000/api/pets', token);
   const MascotasRegistradasData = MascotasRegistradas?.data?.pets;
 
   const MascotasPerdidas = useFetch('http://localhost:8000/api/loss', token);
+
   const MascotasPerdidasData = MascotasPerdidas?.data?.petLoss;
 
   const handleDelete = async (id) => {
-    // Swal.fire({
-    //   title: 'Estas seguro?',
-    //   text: 'No podrás revertir esto!',
-    //   icon: 'warning',
-    //   showCancelButton: true,
-    //   confirmButtonColor: '#3085d6',
-    //   cancelButtonColor: '#d33',
-    //   confirmButtonText: 'Si, borrar',
-    // }).then((result) => {
-    //   if (result.isConfirmed) {
-    //     Swal.fire(
-    //       'Borrado!',
-    //       'Ya no veras esta mascota en tu lista.',
-    //       'success',
-    //     );
-    //   }
-    // });
-
-    axios.delete(`http://localhost:8000/api/pets/${id}`, {
-      headers: { Authorization: `Bearer ${token}` },
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: 'No podrás revertir esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, borrar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`http://localhost:8000/api/pets/${id}`, {
+          headers: { Authorization: `Bearer ${token}` },
+        });
+        Swal.fire(
+          'Borrado!',
+          'Ya no veras esta mascota en tu lista.',
+          'success',
+        );
+      }
     });
   };
   return (
@@ -131,10 +130,10 @@ const UserProfile = () => {
                 <Descripcion>
                      Aquí podrás ver las mascotas que subiste al sitio
                 </Descripcion>
-                <ButtonComponentShort as={'a'} texto={'Agregar otra mascota'} path={'/form-add-pet'}/>
+                <ButtonComponentShort as={'a'} texto={'Agregar una mascota'} path={'/form-add-pet'}/>
             </WapperInfoSection>
           <WrapperListadoCards>
-            {MascotasRegistradas ? (
+            {MascotasRegistradasData?.length !== 0 ? (
               MascotasRegistradasData?.map((mascota) => (
                 <CardMascota
                   key={mascota?.id}
@@ -142,7 +141,7 @@ const UserProfile = () => {
                   nombre={mascota?.name}
                   link={mascota?.images ? mascota?.images[0] : SinFotoMascota}
                   fecha={mascota?.date}
-                  estado={mascota?.estado}
+                  estado={'perdido'}
                   token={token}
                   deleteFunction={() => handleDelete(mascota?.id)}
                 />
@@ -166,15 +165,16 @@ const UserProfile = () => {
             Aquí verás la mascota marcada &quot;perdida&quot;
           </Descripcion>
           <WrapperListadoCards>
-            {MascotasPerdidas ? (
+            {MascotasPerdidasData?.pet !== null ? (
               MascotasPerdidasData?.map((mascota) => (
-                <CardMascota
+                mascota?.pet
+                && <CardMascota
                   key={mascota?.id}
                   id={'/detail-pet'}
-                  nombre={mascota?.name}
-                  link={mascota?.images ? mascota?.images[0] : SinFotoMascota}
+                  nombre={mascota?.pet?.name}
+                  link={mascota?.pet?.images ? mascota?.pet?.images[0] : SinFotoMascota}
                   fecha={mascota?.date}
-                  estado={mascota?.estado}
+                  estado={'perdido'}
                   token={token}
                   deleteFunction={() => handleDelete(mascota?.id)}
                 />
